Made modifications to the Author page. This will later be renamed to about us.
Changed profile picture, background picture and information in the contact section.

Also removed some of the irrelevant sections and made changes to the footer
