+++ src/pages/LandingPages/Author/index.js
@@ -24,7 +24,6 @@ import DefaultNavbar from "examples/Navbars/DefaultNavbar";
 
 // Author page sections
 import Profile from "pages/LandingPages/Author/sections/Profile";
-import Posts from "pages/LandingPages/Author/sections/Posts";
 import Contact from "pages/LandingPages/Author/sections/Contact";
 import Footer from "pages/LandingPages/Author/sections/Footer";
 
@@ -32,7 +31,7 @@ import Footer from "pages/LandingPages/Author/sections/Footer";
 import routes from "routes";
 
 // Images
-import bgImage from "assets/images/city-profile.jpg";
+import bgImage from "assets/images/beach/beach1.jpg";
 
 function Author() {
   return (
@@ -50,14 +49,10 @@ function Author() {
       />
       <MKBox bgColor="white">
         <MKBox
-          minHeight="25rem"
+          minHeight="35rem"
           width="100%"
           sx={{
-            backgroundImage: ({ functions: { linearGradient, rgba }, palette: { gradients } }) =>
-              `${linearGradient(
-                rgba(gradients.dark.main, 0.8),
-                rgba(gradients.dark.state, 0.8)
-              )}, url(${bgImage})`,
+            backgroundImage: `url(${bgImage})`,
             backgroundSize: "cover",
             backgroundPosition: "center",
             display: "grid",
@@ -76,7 +71,6 @@ function Author() {
           }}
         >
           <Profile />
-          <Posts />
         </Card>
         <Contact />
         <Footer />
